3. Int to long

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
 
     public static void main(String[] args) {
-        SalesManager stats = new SalesManager(new int[]{5, -10, 15, -2, 3});
+        SalesManager stats = new SalesManager(new long[]{5, -10, 15, -2, 3});
         System.out.println("Максимальная продажа: " + stats.max());
     }
 }
